+++ src/compiler/getExportedCSS.ts
@@ -22,10 +22,17 @@ export const extractDefaultExport = (exported: unknown) => {
  * Imports default exported CSS from an input file.
  * Resolves esm—commonjs wrappers if neccessary.
  */
-export const getExportedCSS = (filePath: string) =>
-  import(resolve('', filePath)).then(extractDefaultExport).then((css) => {
+export const getExportedCSS = (filePath: string) => {
+  const url = resolve('', filePath);
+  // When `import()`-ing a commonjs modules, simply adding a cacheBust param will
+  // NOTE invalidate the cache. (node 14 and 16 at least)
+  delete require.cache[require.resolve(url)];
+
+  const cacheBust = '?' + Date.now();
+  return import(url + cacheBust).then(extractDefaultExport).then((css) => {
     if (css == null) {
       throw new Error(`Module does not emit string as its default export`);
     }
     return css;
   });
+};